+++ doc/dsl/dsl_zu_level.md
@@ -18,7 +18,7 @@ den DSL-Dateien werden die `level_config`-Definitionen gesucht. Für jede
 `level_config`-Definition kann eine `task_org` Struktur referenzieren, welche mehrere
 `task`-Definition miteinander in Verbindung setzt, hierfür wird ein Ansatz basierend auf
 [Petri-Netzen](../control_mechanisms/petri_nets.md) genutzt, wobei die Nutzereingabe als
-dot-graph erfolgt (“Task-Graph”). Falls nur eine Aufgabe im Level realisiert werden soll,
+dot-taksDependencyGraph erfolgt (“Task-Graph”). Falls nur eine Aufgabe im Level realisiert werden soll,
 kann auch der Zwischenschritt der Task-Organisation übersprungen werden, sodass die
 `task`-Definition direkt in der `level_config` referenziert werden kann.
