+++ dsl/test_resources/program.ds
@@ -1,4 +1,4 @@
-graph g {
+taksDependencyGraph g {
     A -- B
 }
