+++ dungeon/assets/scripts/input1.dng
@@ -1,4 +1,4 @@
-graph g {
+taksDependencyGraph g {
                         G -- T -- Q -- D
                         Q -- F
                         T -- X -- S
