+++ dungeon/src/taksDependencyGraph/TaskDependencyGraph.java
@@ -1,4 +1,4 @@
-package graph;
+package taksDependencyGraph;
 
 import java.util.ArrayList;
 import java.util.Iterator;
